upstream-bug: MicroPython broke path structure
Previously the path for building MicroPython for the ESP32 was a static
path.  Now, the upstream has introduced a variable into the process
which is not needed.  Until I can get upstream fixed, we will need to
patch our build system.

This uses the ${IDF_TARGET} variable in the path for the ESP-IDF
`micropython` component's CMakeLists.txt definition.