components/buses-bug: Change in build requires
At some point one of the include headers needed (`spi_master.h`) was
moved into an ESP-IDF component named `driver`.  This needed to be
added to the build requirements of the `buses` component.

This commit adds the `driver` component to the build requirements of
the `buses` component, ensuring it's done with a private interface.