deprecation: FreeRTOS syntax changed in 2014
In 2014 FreeRTOS v8.0.0 was released which introduced a number of
deprecation notices... some of which it looks like still hadn't been
resolved.

This was discovered via the error message:

> HINT: You maybe using pre FreeRTOS V8.0.0 data types. The backward
compatibility of such data types is no longer enabled by default. Please
turn on CONFIG_FREERTOS_ENABLE_BACKWARD_COMPATIBILITY explicitly to use
such data types.

I was about to just flip that switch in the config and move on... until
I saw espressif/esp-idf#51.

Yes... Issue number 51.  In a repository which currently is working
through issue number 11833, number 51 debated what I was just talking
about doing... but in 2016.

Obviously this is a band-aid which needed to be ripped off (and next up
will need to be `driver_mch22`.  The fix is simple enough.
`portTICK_RATE_MS` needs to be replaced with `portTICK_PERIOD_MS`.  This
can be a simple `sed` replacement:

    sed  -i 's/portTICK_RATE_MS/portTICK_PERIOD_MS/g' components/buses/buses.c