xtensa: GCC 11 type int32_t changes int -> long
As documented in the [ESP-IDF][1]:

> The types `int32_t` and `uint32_t` have been changed from the previous
`int` and `unsigned int` to `long` and `unsigned long` respectively for
the Xtensa compiler. This change now matches upstream GCC which `long`
integers for `int32_t` and `uint32_t` on Xtensa, RISC-V, and other
architectures.

This changes the type of the pointer passed to nvs_flash_init() from
(int) to (long).

[1]: https://docs.espressif.com/projects/esp-idf/en/latest/esp32/migration-guides/release-5.x/5.0/gcc.html#int32-t-and-uint32-t-for-xtensa-compiler

bug: missing header file

This header file was missing from the migration of the micropython
component?